+++ src/main/java/com/younghun/hibusgo/domain/Region.java
@@ -13,10 +13,10 @@ import lombok.ToString;
 @ToString
 public class Region {
 
-    //지역 아이디
-    private int id;
+    // 지역 아이디
+    private long id;
 
-    //지역 이름
+    // 지역 이름
     private String name;
 
     // 지역 상태(기본, 삭제)
